groupe 7/app/mavoiture/mavoiture.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { MavoitureComponent } from './mavoiture.component';

describe('Component: Mavoiture', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [MavoitureComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([MavoitureComponent],
      (component: MavoitureComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(MavoitureComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(MavoitureComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <yaka-mavoiture></yaka-mavoiture>
  `,
  directives: [MavoitureComponent]
})
class MavoitureComponentTestController {
}

